protocol: hoist BaudRate from osc-servo-core into the wire crate
